+++ src/methods/object/multipart/completeMultipartUpload.ts
@@ -1,3 +1,4 @@
+import TosClientError from '../../../TosClientError';
 import TOSBase from '../../base';
 
 export interface CompleteMultipartUploadInput {
@@ -8,7 +9,16 @@ export interface CompleteMultipartUploadInput {
     eTag: string;
     partNumber: number;
   }[];
+  /**
+   * when true `parts` param need to be empty array
+   */
+  completeAll?: boolean;
 }
+
+export type UploadedPart = {
+  PartNumber: number;
+  ETag: string;
+};
 
 export interface CompleteMultipartUploadOutput {
   Bucket: string;
@@ -17,12 +27,41 @@ export interface CompleteMultipartUploadOutput {
   Location: string;
   VersionID?: string;
   HashCrc64ecma?: number;
+  /** the field has a value when completeAll is true */
+  CompletedParts?: UploadedPart[];
 }
 
 export async function completeMultipartUpload(
   this: TOSBase,
   input: CompleteMultipartUploadInput
 ) {
+  if (input.completeAll) {
+    if (input.parts?.length > 0) {
+      throw new TosClientError(
+        `Should not specify both 'completeAll' and 'parts' params.`
+      );
+    }
+    return this.fetchObject<CompleteMultipartUploadOutput>(
+      input,
+      'POST',
+      {
+        uploadId: input.uploadId,
+      },
+      {
+        'x-tos-complete-all': 'yes',
+      },
+      undefined,
+      {
+        handleResponse(response) {
+          return {
+            ...response.data,
+            VersionID: response.headers['x-tos-version-id'],
+          };
+        },
+      }
+    );
+  }
+
   return this.fetchObject<CompleteMultipartUploadOutput>(
     input,
     'POST',